refactor(resume): extract hash function to separate module
The hashCode function has been moved out of the experiences.tsx into a
separate utility file generate-djb2-hash.ts. Corresponding unit tests
were added in generate-djb2-hash.spec.ts to ensure the correctness of
the hashing functionality.
